# main.js
const express = require('express');
const { program } = require('commander');
const fs = require('fs').promises;
const fss = require('fs');
const path = require('path');
const multer = require('multer');
const swaggerUi = require('swagger-ui-express');
const YAML = require('yamljs');

// налаштування CLI

program
  .requiredOption('-H, --host <host>', 'адреса сервера')
  .requiredOption('-p, --port <port>', 'порт сервера')
  .requiredOption('-c, --cache <path>', 'шлях до директорії з кешом');

program.parse(process.argv);
const options = program.opts();
const host = options.host;
const port = options.port;
const cacheDir = path.resolve(options.cache);

// перевірка папки кешу

if (!fss.existsSync(cacheDir)) {
  console.log(`Директорія кешу не існує. Створюємо: ${cacheDir}`);
  try {
      fss.mkdirSync(cacheDir, { recursive: true });
  } catch (err) {
      console.error(`Помилка створення директорії: ${err.message}`);
      process.exit(1);
  }
} else {
  console.log(`Директорія кешу існує: ${cacheDir}`);
}

// масив для збереження даних

let inventory = []; 

//налаштування multer для завантаження фото

const storage = multer.diskStorage({
  destination: function (req, file, cb) {
    cb(null, cacheDir);
  },
  filename: function (req, file, cb) {
    cb(null, file.originalname);
  }
});

const upload = multer({ storage: storage });
const app = express();



app.use(express.json());
app.use(express.urlencoded({ extended: true }));

//  документація 

const swaggerDocument = YAML.load('./swagger.yaml');
app.use('/docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));

// html форми

app.get('/RegisterForm.html', (req, res) => {
    res.sendFile(path.join(__dirname, 'RegisterForm.html'));
});

app.get('/SearchForm.html', (req, res) => {
    res.sendFile(path.join(__dirname, 'SearchForm.html'));
});

/**
 * GET /inventory
 * Отримання списку всіх інвентаризованих речей.
 * Повертає масив JSON об'єктів.
 */

app.get('/inventory', (req, res) => {
    res.status(200).json(inventory);
});

/**
 * GET /inventory/:id
 * Отримання інформації про конкретну річ за ID.
 * Повертає JSON об'єкт або 404.
 */

app.get('/inventory/:id', (req, res) => {
    const item = inventory.find(i => i.id === req.params.id);
    if (!item) return res.status(404).send('Not Found');
    res.json(item);
});

/**
 * GET /inventory/:id/photo
 * Отримання фото зображення конкретної речі.
 * Віддає файл зображення (image/jpeg).
 */

app.get('/inventory/:id/photo', (req, res) => {
    const item = inventory.find(i => i.id === req.params.id);
    if (!item || !item.photo) return res.status(404).send('Not Found');
    
    const photoPath = path.join(cacheDir, item.photo);
    if (!fss.existsSync(photoPath)) return res.status(404).send('Photo file missing');
    
    res.sendFile(photoPath);
});

/**
 * POST /register
 * Реєстрація нового пристрою.
 * Використовує multipart/form-data. Приймає ім'я, опис та фото.
 */

app.post('/register', upload.single('photo'), (req, res) => {
    const { inventory_name, description } = req.body;
    if (!inventory_name) return res.status(400).send('Bad Request: inventory_name is required');

    const newItem = {
        id: Date.now().toString(),
        name: inventory_name,
        description: description || '',
        photo: req.file ? req.file.originalname : null
    };

    inventory.push(newItem);
    console.log(`Додано: ${newItem.name} (ID: ${newItem.id})`);
    res.status(201).send(`Item created with ID: ${newItem.id}`);
});

/**
 * POST /search
 * Пошук пристрою за ID.
 * Використовує x-www-form-urlencoded.
 * Параметр has_photo додає посилання на фото в опис.
 */

app.post('/search', (req, res) => {
    const { id, has_photo } = req.body;
    const item = inventory.find(i => i.id === id);
    if (!item) return res.status(404).send('Not Found');
    
    const responseItem = { ...item };
    if (has_photo === 'yes' && item.photo) {
        responseItem.description += ` (Посилання на фото: /inventory/${item.id}/photo)`;
    }
    res.status(200).json(responseItem);
});


/**
 * PUT /inventory/:id
 * Оновлення імені або опису конкретної речі.
 * Приймає JSON.
 */

app.put('/inventory/:id', (req, res) => {
    const item = inventory.find(i => i.id === req.params.id);
    if (!item) return res.status(404).send('Not Found');

    const { name, description } = req.body;
    

    if (name) item.name = name;
    if (description) item.description = description;

    res.json(item);
});


/**
 * PUT /inventory/:id/photo
 * Оновлення фото зображення конкретної речі.
 * Приймає файл через multipart/form-data.
 */

app.put('/inventory/:id/photo', upload.single('photo'), (req, res) => {
    const item = inventory.find(i => i.id === req.params.id);
    if (!item) return res.status(404).send('Not Found');

    if (!req.file) return res.status(400).send('No file uploaded');


    item.photo = req.file.originalname;
    
    res.send('Photo updated');
});


/**
 * DELETE /inventory/:id
 * Видалення інвентаризованої речі зі списку.
 */

app.delete('/inventory/:id', (req, res) => {
    const index = inventory.findIndex(i => i.id === req.params.id);
    if (index === -1) return res.status(404).send('Not Found');


    inventory.splice(index, 1);
    
    res.send('Item deleted');
});

// Обробка неправильних методів (405)

app.all(/^\/inventory/, (req, res) => {
     res.status(405).send('Method Not Allowed');
});


// головна сторінка



app.get('/', (req, res) => {
    res.send(`
        <h1>Сервіс Інвентаризації</h1>
        <ul>
           <li><a href="/docs">Swagger Документація</a></li>
            <li><a href="/RegisterForm.html">Зареєструвати річ (HTML форма)</a></li>
            <li><a href="/SearchForm.html">Знайти річ (HTML форма)</a></li>
            <li><a href="/inventory">Список всіх речей (JSON)</a></li>
        </ul>
    `);
});


app.listen(port, host, () => {
  console.log(`Сервер запущено на http://${host}:${port}`);
});